add DTO design for roles

+++ src/main/java/org/green_building/excellent_training/services/RolesService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import org.green_building.excellent_training.entities.Role;
+import org.green_building.excellent_training.dtos.RoleDto;
 import org.green_building.excellent_training.repositories.RolesRepository;
 
 @Service
@@ -17,36 +18,43 @@ public class RolesService {
         this.rolesRepository = rolesRepository;
     }
 
-    public List<Role> getAll() {
-        return this.rolesRepository.findAll();
+    public List<RoleDto> getAll() {
+        List<Role> roles = this.rolesRepository.findAll();
+        return RoleDto.from(roles);
     }
 
-    public Role getById(Integer id) {
+    public RoleDto getById(Integer id) {
         // findById(id) returns Optional<Role>
-        return this.rolesRepository.findById(id).orElse(null);
+        Role role = this.rolesRepository.findById(id).orElse(null);
+        return RoleDto.from(role);
     }
 
-    public Role create(Role role) {
-        if (role.getName() == null) return null;
-        return this.rolesRepository.save(role);
+    public RoleDto create(RoleDto roleDto) {
+        if (roleDto.getName() == null/* || this.rolesRepository.findByName(roleDto.getName()).isPresent()*/) return null; // TODO: check this commented line, in case in the future you want to check for unique values(if you will then do that across all the projects) (if ou will do so then make sure to give back useful response errors to the frontend in case of a fail)
+        Role role = Role.from(roleDto);
+        Role createdRole = this.rolesRepository.save(role);
+        return RoleDto.from(createdRole);
     }
 
-    public Role updateById(Integer id, Role updates) {
-        Role role = this.getById(id);
+    public RoleDto updateById(Integer id, RoleDto updatesDto) {
+        Role updates = Role.from(updatesDto);
+        Role role = this.rolesRepository.findById(id).orElse(null);
         if (role == null) return null;
         if (updates.getName() != null) role.setName(updates.getName());
-        return this.rolesRepository.save(role);
+        Role updatedRole = this.rolesRepository.save(role);
+        return RoleDto.from(updatedRole);
     }
 
-    public List<Role> deleteAll() {
+    public List<RoleDto> deleteAll() {
         List<Role> roles = this.rolesRepository.findAll();
         this.rolesRepository.deleteAll();
-        return roles;
+        return RoleDto.from(roles);
     }
 
-    public Role deleteById(Integer id) {
-        Role role = this.getById(id); // if it doesn't exist, then this will be null
+    public RoleDto deleteById(Integer id) {
+        RoleDto roleDto = this.getById(id); // if it doesn't exist, then this will be null
+        if (roleDto == null) return null;
         this.rolesRepository.deleteById(id);
-        return role;
+        return roleDto;
     }
 }
